ros hid tx/rx compiles

+++ lunabot_embedded/firmware/teensy_main/lib/lunabot_drivers/robot.hpp
@@ -20,25 +20,25 @@ extern Sabertooth MC4; // high-current
 namespace actuation {
 void cb(int8_t lead_screw, int8_t lin_act);
 void loop_once();
-void update(int16_t *, int16_t *);
+void update(int32_t *, int32_t *);
 } // namespace actuation
 
 namespace drivetrain {
 void cb(int8_t left, int8_t right);
 void loop_once();
-void update(int16_t *, int16_t *);
+void update(int32_t *, int32_t *);
 } // namespace drivetrain
 
 namespace deposition {
 void cb(int8_t dep);
 void loop_once();
-void update(int16_t *);
+void update(int32_t *);
 } // namespace deposition
 
 namespace excavation {
 void cb(int8_t exc);
 void loop_once();
-void update(int16_t *);
+void update(int32_t *);
 } // namespace excavation
 
 #endif
